backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# HTTP Client
httpx==0.25.2
aiofiles==22.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pyyaml==6.0.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging
structlog==23.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# AI and ML (Optional - for future enhancements)
# transformers==4.36.0
# torch==2.1.1
# scikit-learn==1.3.2